feat(database): 集成 MariaDB 存储 mr 数据

+++ src/main/java/com/luke/sirius/MRServer/data/GitlabWebhookData.java
@@ -72,7 +72,7 @@ public class GitlabWebhookData {
         public Boolean draft;
         public Integer head_pipeline_id;
         public Integer id;
-        public Integer iid;
+        public Long iid;
         public String merge_status;
         public String source_branch;
         public Integer source_project_id;
